Ajustes

+++ Dominio/Questionarios/Entidades/Questionario.cs
@@ -8,7 +8,7 @@ namespace Dominio.Questionarios.Entidades
         public virtual int Id { get; protected set; }
         public virtual DateTime? DataInicio { get; protected set; }
         public virtual StatusQuestionarioEnum Status { get; protected set; }
-        public virtual int Porcentagem { get; protected set; }
+        public virtual double Porcentagem { get; protected set; }
         public virtual DateTime? DataConclusao { get; protected set; }
         public virtual IList<QuestionarioQuestao> Questoes { get; protected set; } = new List<QuestionarioQuestao>();
 
@@ -20,7 +20,7 @@ namespace Dominio.Questionarios.Entidades
         public virtual void SetStatus(StatusQuestionarioEnum Status){ 
             this.Status = Status;
         }
-        public virtual void SetPorcentagem(int porcentagem) {  
+        public virtual void SetPorcentagem(double porcentagem) {  
             this.Porcentagem = porcentagem;
         }
         public virtual void SetDataConclusao(DateTime? dataConclusao) {
